Remove Features details...
Remove the details of the Features epic and just link to the
master list for shopping list features. Mention features that are
not yet implemented.

Expand on "Using the app" a little bit more. The app is easy
enough to use w/o instructions, but the README should explain
enough to give a good idea of what it does.